Update to add Cycle for grouping

+++ scripts/import.js
@@ -2,6 +2,8 @@ var axios = require("axios");
 var path = require("path");
 var fs = require("fs");
 
+const DRY_RUN = false;
+
 const doImport = async () => {
   const rootDir = path.join(__dirname, "..");
   const packDir = path.join(__dirname, "..", "packs");
@@ -23,12 +25,14 @@ const doImport = async () => {
   const hob_cookies =
     "DefaultSort=SortPopularity; ProductFilter=ProductAll; OwnedProducts=; SetSearch=SearchCommunity";
 
+  const hobSetsByName = {};
   const hobSets = await axios.get(hob_allSetsUrl, {
     headers: { Cookie: hob_cookies },
   });
 
-  for (let pack of hobSets.data) {
+  for (let pack of hobSets.data) {    
     console.log(`Working with pack ${pack.Name}`);
+    hobSetsByName[pack.Name.toLowerCase()] = pack;
     // get all the player cards for the pack
     const cardsUrl = `${hallOfBeorn}?CardSet=${encodeURIComponent(
       pack.Name
@@ -50,36 +54,53 @@ const doImport = async () => {
     headers: { Cookie: hob_cookies },
   });
 
-  fs.writeFileSync(
-    path.join(rootDir, "scenarios.json"),
-    JSON.stringify(hobScenarios.data, null, 4)
-  );
-
   for (let scenario of hobScenarios.data) {
-    console.log(`Working with scenario ${scenario.Title}`);
+    if (scenario.Title.indexOf('Rhosgobel') !== -1) {
+      console.log(`Working with scenario ${scenario.Title}`);
+    }
+
+    //Update the scenario to include Cycle
+    scenario.Cycle =
+      hobSetsByName[scenario.Title.toLowerCase()]?.Cycle ||
+      hobSetsByName[scenario.Product.toLowerCase()]?.Cycle;
+
     const scenCardsUrl = `${hallOfBeorn}/?Scenario=${encodeURIComponent(
       scenario.Title
     )}`;
-    console.log(`\tgetting the scenario`);
+    console.log(`\tgetting the scenario ${scenario.Title}`);
     const sc = await axios.get(scenCardsUrl, {
       headers: { Cookie: hob_cookies },
     });
     console.log(`\tgot scenario. Saving json...`);
-    fs.writeFileSync(
-      path.join(scenariosDir, scenario.Title + ".json"),
-      JSON.stringify(
-        {
-          Title: scenario.Title,
-          Slug: scenario.Slug,
-          Product: scenario.Product,
-          Number: scenario.Number,
-          AllCards: sc.data,
-        },
-        null,
-        4
-      )
-    );
+    if (DRY_RUN) {
+      console.log(scenario);
+    } else {
+      fs.writeFileSync(
+        path.join(scenariosDir, scenario.Title + ".json"),
+        JSON.stringify(
+          {
+            Title: scenario.Title,
+            Slug: scenario.Slug,
+            Product: scenario.Product,
+            Number: scenario.Number,
+            AllCards: sc.data,
+            Cycle: scenario.Cycle,
+          },
+          null,
+          4
+        )
+      );
+    }
     console.log("\tSaved json.");
+  }
+
+  if (DRY_RUN) {
+    // console.log(hobScenarios.data);
+  } else {
+    fs.writeFileSync(
+      path.join(rootDir, "scenarios.json"),
+      JSON.stringify(hobScenarios.data, null, 4)
+    );
   }
 };
